VentaNegocio
Se comienza a dar funcionalidad a VentaNegocio + btnConfirmarCompra

+++ Dominio/Venta.cs
@@ -11,7 +11,7 @@ namespace Dominio
         public int IDUsuario { get; set; }
         public char FormaPago { get; set; }
         public char MetodoEnvio { get; set; }
-        public int PrecioTot { get; set; }
+        public decimal PrecioTot { get; set; }
         public int CantTot { get; set; }
         public Direccion DomicilioEntrega { get; set; }
         public DateTime Fecha { get; set; }
